+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/referee_config.js
@@ -135,7 +135,8 @@ define([
         inventory,
         playerTag,
         refereeCoop.getOrderedSubcommanders(inventory, game).length,
-        battleRng
+        battleRng,
+        { ffa: !_.isEmpty(ai.foes) }
       );
     }
 
@@ -168,7 +169,7 @@ define([
       bounty_mode:
         ai.bountyMode ||
         gwoCards.anyPlayerHasCard(inventory, "gwaio_enable_bounties", game),
-      bounty_value: ai.bountyModeValue,
+      bounty_value: gwoAI.bountyValue(ai),
       sudden_death_mode:
         ai.suddenDeath ||
         gwoCards.anyPlayerHasCard(inventory, "gwaio_enable_suddendeath", game),